Driver
This will be the main game engine that controls Reversi (Othello)
So far basic piece placement has been implemented. This includes
some checks for in bounds and placement validity. Pointers and
dynamic list / linked lists have been used to increase efficiency
of the program when checking for and flipping pieces.